refactor: main.h->minirt.h변경 및 인클루드 헤더 위치 변경 리팩토링

+++ include/get_next_line.h
@@ -6,16 +6,18 @@
 /*   By: junyoung <junyoung@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/09/10 12:37:40 by junyoung          #+#    #+#             */
-/*   Updated: 2023/05/13 12:02:33 by junyoung         ###   ########seoul.kr  */
+/*   Updated: 2023/06/01 20:20:59 by junyoung         ###   ########seoul.kr  */
 /*                                                                            */
 /* ************************************************************************** */
 
 #ifndef GET_NEXT_LINE_H
 # define GET_NEXT_LINE_H
 
+# include <stdlib.h>
+
 char	*get_next_line(int fd);
 
-size_t	ft_strlen2(const char *str);
+int		ft_strlen2(const char *str);
 char	*ft_strdup2(char *src);
 char	*ft_strjoin(char *s1, char *s2);
 int		ft_strchr(const char *str, int c);
